Added a unit test which covers the average calculated latency. Updated based on comments.
